pyproject.toml
[tool.poetry]
name = "authflaskapp"
version = "0.1.0"
description = ""
authors = ["zikrulla <zikrullayashinov173@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
passlib = "^1.7.4"
flask-jwt-extended = "^4.6.0"
flask-marshmallow = "^1.2.0"
marshmallow-sqlalchemy = "^1.0.0"
flask-query-builder = "^0.3.0"
flask-restful = "^0.3.10"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
